src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.24.0",
    "azure-storage-blob>=12.26.0",
    "load-dotenv>=0.1.0",
    "openai>=1.99.7",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
